python-infer/app/requirements.txt
# FastAPI OCR Multi-Pipeline API 依赖配置

# FastAPI核心
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# OCR核心 (PaddleOCR 3.x)
paddleocr[all]
# paddlepaddle-gpu 需单独安装（requirements.txt不支持-i参数）:
# python -m pip install paddlepaddle-gpu==3.2.0 -i https://www.paddlepaddle.org.cn/packages/stable/cu126/

# 图像处理
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0

# HTTP客户端
requests>=2.31.0
httpx>=0.25.0

# 日志和监控
python-json-logger>=2.0.7
